# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Movie recommendation service for the PRPO project",
        "title": "Predlogi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/predlogi",
    "paths": {
        "/api/v1/predlogi/admin/trigger-job": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Triggers the recommendation generation process for all users",
                "tags": [
                    "admin"
                ],
                "summary": "Manually trigger recommendation generation job",
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middleware.HttpError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/middleware.HttpError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.HttpError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "middleware.HttpError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}